refactor: split collection.svelte.ts into collection/navigation/windowLayout stores (#611)
Splits the collection.svelte.ts god-object store (1,331 lines mixing
library data, navigation history, and window/miniplayer geometry) into
three focused stores: collectionStore (trimmed to library/search/scan
data), navigationStore (active tab/selection/back-forward history), and
windowLayoutStore (sidebar/panel/immersive/miniplayer/viewport state).
Updates every consuming component's imports accordingly.

Part of the #577 code-quality roadmap, item 21.
