+++ src/lib/components/PlaylistsCollectionView.svelte
@@ -3,6 +3,7 @@
   import { onMount } from "svelte";
   import { open } from "@tauri-apps/plugin-dialog";
   import { collectionStore } from "../stores/collection.svelte";
+  import { navigationStore } from "../stores/navigation.svelte";
   import { playlistsStore } from "../stores/playlists.svelte";
   import { playerStore } from "../stores/player.svelte";
   import { i18n } from "../stores/i18n.svelte";
@@ -233,11 +234,11 @@
   });
 
   let activeViewMode = $derived(
-    collectionStore.playlistsSubTab === "auto" ? prefs.playlistsAutoViewMode : prefs.playlistsCustomViewMode
+    navigationStore.playlistsSubTab === "auto" ? prefs.playlistsAutoViewMode : prefs.playlistsCustomViewMode
   );
 
   function setActiveViewMode(mode: CollectionViewMode) {
-    if (collectionStore.playlistsSubTab === "auto") {
+    if (navigationStore.playlistsSubTab === "auto") {
       prefs.setPlaylistsAutoViewMode(mode);
     } else {
       prefs.setPlaylistsCustomViewMode(mode);
@@ -245,7 +246,7 @@
   }
 
   function openAuto(def: AutoDef) {
-    collectionStore.viewAutoPlaylist(
+    navigationStore.viewAutoPlaylist(
       def.kind === "genre"
         ? { kind: "genre", genre: def.genre, playlistId: def.playlistId, updated: def.updated }
         : def.kind === "decade"
@@ -258,14 +259,14 @@
 
   function openPlaylist(pl: Playlist) {
     playlistsStore.selectPlaylist(pl.id);
-    collectionStore.viewPlaylist(pl.id);
+    navigationStore.viewPlaylist(pl.id);
   }
 
   async function handleCreateBlankPlaylist() {
     try {
       const playlist = await playlistsStore.createPlaylist(i18n.t("playlists.untitledPlaylistName"));
       if (playlist) {
-        collectionStore.viewPlaylist(playlist.id);
+        navigationStore.viewPlaylist(playlist.id);
       }
     } catch (err) {
       console.error("Failed to create playlist:", err);
@@ -282,7 +283,7 @@
       if (selected && typeof selected === "string") {
         await playlistsStore.importPlaylist(selected);
         if (playlistsStore.activePlaylistId !== null) {
-          collectionStore.viewPlaylist(playlistsStore.activePlaylistId);
+          navigationStore.viewPlaylist(playlistsStore.activePlaylistId);
         }
       }
     } catch (err) {
@@ -291,15 +292,15 @@
   }
 </script>
 
-{#if collectionStore.selectedPlaylistId !== null}
+{#if navigationStore.selectedPlaylistId !== null}
   <PlaylistView />
-{:else if collectionStore.selectedAutoPlaylist !== null}
-  <AutoPlaylistDetailView view={collectionStore.selectedAutoPlaylist} />
+{:else if navigationStore.selectedAutoPlaylist !== null}
+  <AutoPlaylistDetailView view={navigationStore.selectedAutoPlaylist} />
 {:else}
   <div class="flex-1 flex flex-col overflow-hidden bg-brand-main text-brand-text-secondary h-full">
-    <div class="flex-1 px-6 overflow-y-auto {playerStore.currentSong ? 'pb-28' : 'pb-6'}" use:rememberScroll={`playlists:${collectionStore.playlistsSubTab}`}>
+    <div class="flex-1 px-6 overflow-y-auto {playerStore.currentSong ? 'pb-28' : 'pb-6'}" use:rememberScroll={`playlists:${navigationStore.playlistsSubTab}`}>
       <div class="sticky top-0 z-20 bg-brand-main pt-3">
-        {#if collectionStore.playlistsSubTab === "custom"}
+        {#if navigationStore.playlistsSubTab === "custom"}
           <div class="h-10 flex items-center gap-2 mb-2">
             <Button onclick={handleCreateBlankPlaylist} variant="primary" title={i18n.t('playlists.newPlaylistBtn')}>
               <Plus class="w-4 h-4" />
@@ -318,7 +319,7 @@
 
         <div class="h-12 flex items-center justify-between">
           <div class="text-xs text-brand-text-secondary font-medium">
-            {#if collectionStore.playlistsSubTab === "auto"}
+            {#if navigationStore.playlistsSubTab === "auto"}
               {sortedAutoDefs.length === 1 ? i18n.t('playlists.showingOnePlaylist') : i18n.t('playlists.showingPlaylists', { count: sortedAutoDefs.length })}
             {:else}
               {sortedPlaylists.length === 1 ? i18n.t('playlists.showingOnePlaylist') : i18n.t('playlists.showingPlaylists', { count: sortedPlaylists.length })}
@@ -356,7 +357,7 @@
               </button>
             </div>
             <div class="relative">
-            {#if collectionStore.playlistsSubTab === "auto"}
+            {#if navigationStore.playlistsSubTab === "auto"}
               <Select
                 value={`${autoSortField}-${autoSortAsc}`}
                 onchange={(e) => {
@@ -397,7 +398,7 @@
       </div>
 
       <div class="pt-2 pb-8">
-        {#if collectionStore.playlistsSubTab === "auto"}
+        {#if navigationStore.playlistsSubTab === "auto"}
           <div class="grid {activeViewMode === 'rows' ? 'grid-cols-[repeat(auto-fill,minmax(280px,1fr))] gap-2' : 'grid-cols-[repeat(auto-fill,minmax(180px,1fr))] gap-5'}">
             {#each sortedAutoDefs as def (def.id)}
               {#if activeViewMode === "rows"}
