+++ src/lib/components/Sidebar.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
   import { collectionStore } from "../stores/collection.svelte";
+  import { navigationStore } from "../stores/navigation.svelte";
   import { playlistsStore } from "../stores/playlists.svelte";
   import { playerStore } from "../stores/player.svelte";
   import { themeStore } from "../stores/theme.svelte";
@@ -21,19 +22,19 @@
   let showUpdateBadge = $derived(updaterStore.updateAvailable || updaterStore.installStatus === "ready-to-restart");
 
   function selectCollectionTab() {
-    collectionStore.activeTab = "collection";
-    collectionStore.selectedArtistName = null;
-    collectionStore.selectedAlbumName = null;
+    navigationStore.activeTab = "collection";
+    navigationStore.selectedArtistName = null;
+    navigationStore.selectedAlbumName = null;
   }
 
   function selectPlaylistsTab() {
-    collectionStore.activeTab = "playlists";
-    collectionStore.selectedPlaylistId = null;
-    collectionStore.selectedAutoPlaylist = null;
+    navigationStore.activeTab = "playlists";
+    navigationStore.selectedPlaylistId = null;
+    navigationStore.selectedAutoPlaylist = null;
   }
 
   function navigateToFoldersSettings() {
-    collectionStore.activeTab = "settings";
+    navigationStore.activeTab = "settings";
     invoke("set_app_setting", { key: "active_settings_tab", value: "folders" });
   }
 
@@ -53,10 +54,10 @@
   }
 
   function openPlaylistsSubTab(subTab: "auto" | "custom") {
-    collectionStore.activeTab = "playlists";
-    collectionStore.playlistsSubTab = subTab;
-    collectionStore.selectedPlaylistId = null;
-    collectionStore.selectedAutoPlaylist = null;
+    navigationStore.activeTab = "playlists";
+    navigationStore.playlistsSubTab = subTab;
+    navigationStore.selectedPlaylistId = null;
+    navigationStore.selectedAutoPlaylist = null;
   }
 
 </script>
@@ -64,8 +65,8 @@
 <aside style="width: {width}px;" class="bg-brand-sidebar flex flex-col h-full text-brand-text-secondary select-none flex-shrink-0 overflow-hidden transition-[width] duration-200 ease-out {themeStore.isGlassTheme ? 'glass-surface' : ''}" class:transition-none={resizing}>
   <nav class="{isCollapsed ? 'p-2' : 'p-4'} space-y-0.5 flex flex-col items-center">
     <button
-      onclick={() => { collectionStore.activeTab = "home"; }}
-      class="flex items-center gap-3 transition-all duration-150 {collectionStore.activeTab === 'home' ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
+      onclick={() => { navigationStore.activeTab = "home"; }}
+      class="flex items-center gap-3 transition-all duration-150 {navigationStore.activeTab === 'home' ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
       title={i18n.t('sidebar.home')}
     >
       <Home class={isCollapsed ? "w-5 h-5" : "w-4 h-4"} />
@@ -78,16 +79,16 @@
     <div class="w-full flex flex-col {isCollapsed ? 'items-center' : ''}">
       <button
         onclick={selectCollectionTab}
-        class="flex items-center gap-3 transition-all duration-150 {collectionStore.activeTab === 'collection' && isCollapsed ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
+        class="flex items-center gap-3 transition-all duration-150 {navigationStore.activeTab === 'collection' && isCollapsed ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
         title={i18n.t('sidebar.collection')}
       >
-        {#if isCollapsed && collectionStore.activeTab === 'collection' && collectionStore.activeSubTab === 'artists'}
+        {#if isCollapsed && navigationStore.activeTab === 'collection' && navigationStore.activeSubTab === 'artists'}
           <Mic2 class="w-5 h-5" />
-        {:else if isCollapsed && collectionStore.activeTab === 'collection' && collectionStore.activeSubTab === 'albums'}
+        {:else if isCollapsed && navigationStore.activeTab === 'collection' && navigationStore.activeSubTab === 'albums'}
           <DiscAlbum class="w-5 h-5" />
-        {:else if isCollapsed && collectionStore.activeTab === 'collection' && collectionStore.activeSubTab === 'songs'}
+        {:else if isCollapsed && navigationStore.activeTab === 'collection' && navigationStore.activeSubTab === 'songs'}
           <Music class="w-5 h-5" />
-        {:else if isCollapsed && collectionStore.activeTab === 'collection' && collectionStore.activeSubTab === 'genres'}
+        {:else if isCollapsed && navigationStore.activeTab === 'collection' && navigationStore.activeSubTab === 'genres'}
           <Tag class="w-5 h-5" />
         {:else}
           <Library class={isCollapsed ? "w-5 h-5" : "w-4 h-4"} />
@@ -100,8 +101,8 @@
       {#if !isCollapsed}
         <div class="pl-4 pr-1 py-0.5 space-y-0 border-l-2 border-brand-accent/30 ml-4 my-0.5">
           <button
-            onclick={() => { collectionStore.activeTab = "collection"; collectionStore.activeSubTab = "artists"; collectionStore.selectedArtistName = null; collectionStore.selectedAlbumName = null; }}
-            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {collectionStore.activeTab === 'collection' && collectionStore.activeSubTab === 'artists' && !collectionStore.selectedArtistName && !collectionStore.selectedAlbumName ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
+            onclick={() => { navigationStore.activeTab = "collection"; navigationStore.activeSubTab = "artists"; navigationStore.selectedArtistName = null; navigationStore.selectedAlbumName = null; }}
+            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {navigationStore.activeTab === 'collection' && navigationStore.activeSubTab === 'artists' && !navigationStore.selectedArtistName && !navigationStore.selectedAlbumName ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
           >
             <div class="flex items-center gap-2 truncate">
               <Mic2 class="w-3.5 h-3.5" />
@@ -113,8 +114,8 @@
           </button>
 
           <button
-            onclick={() => { collectionStore.activeTab = "collection"; collectionStore.activeSubTab = "albums"; collectionStore.selectedArtistName = null; collectionStore.selectedAlbumName = null; }}
-            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {collectionStore.activeTab === 'collection' && collectionStore.activeSubTab === 'albums' && !collectionStore.selectedArtistName && !collectionStore.selectedAlbumName ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
+            onclick={() => { navigationStore.activeTab = "collection"; navigationStore.activeSubTab = "albums"; navigationStore.selectedArtistName = null; navigationStore.selectedAlbumName = null; }}
+            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {navigationStore.activeTab === 'collection' && navigationStore.activeSubTab === 'albums' && !navigationStore.selectedArtistName && !navigationStore.selectedAlbumName ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
           >
             <div class="flex items-center gap-2 truncate">
               <DiscAlbum class="w-3.5 h-3.5" />
@@ -126,8 +127,8 @@
           </button>
 
           <button
-            onclick={() => { collectionStore.activeTab = "collection"; collectionStore.activeSubTab = "songs"; collectionStore.selectedArtistName = null; collectionStore.selectedAlbumName = null; }}
-            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {collectionStore.activeTab === 'collection' && collectionStore.activeSubTab === 'songs' && !collectionStore.selectedArtistName && !collectionStore.selectedAlbumName ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
+            onclick={() => { navigationStore.activeTab = "collection"; navigationStore.activeSubTab = "songs"; navigationStore.selectedArtistName = null; navigationStore.selectedAlbumName = null; }}
+            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {navigationStore.activeTab === 'collection' && navigationStore.activeSubTab === 'songs' && !navigationStore.selectedArtistName && !navigationStore.selectedAlbumName ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
           >
             <div class="flex items-center gap-2 truncate">
               <Music class="w-3.5 h-3.5" />
@@ -139,8 +140,8 @@
           </button>
 
           <button
-            onclick={() => { collectionStore.activeTab = "collection"; collectionStore.activeSubTab = "genres"; collectionStore.selectedArtistName = null; collectionStore.selectedAlbumName = null; }}
-            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {collectionStore.activeTab === 'collection' && collectionStore.activeSubTab === 'genres' && !collectionStore.selectedArtistName && !collectionStore.selectedAlbumName ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
+            onclick={() => { navigationStore.activeTab = "collection"; navigationStore.activeSubTab = "genres"; navigationStore.selectedArtistName = null; navigationStore.selectedAlbumName = null; }}
+            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {navigationStore.activeTab === 'collection' && navigationStore.activeSubTab === 'genres' && !navigationStore.selectedArtistName && !navigationStore.selectedAlbumName ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
           >
             <div class="flex items-center gap-2 truncate">
               <Tag class="w-3.5 h-3.5" />
@@ -157,10 +158,10 @@
     <div class="w-full flex flex-col {isCollapsed ? 'items-center' : ''}">
       <button
         onclick={selectPlaylistsTab}
-        class="flex items-center gap-3 transition-all duration-150 {collectionStore.activeTab === 'playlists' && isCollapsed ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
+        class="flex items-center gap-3 transition-all duration-150 {navigationStore.activeTab === 'playlists' && isCollapsed ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
         title={i18n.t('sidebar.playlists')}
       >
-        {#if isCollapsed && collectionStore.activeTab === 'playlists' && collectionStore.playlistsSubTab === 'auto'}
+        {#if isCollapsed && navigationStore.activeTab === 'playlists' && navigationStore.playlistsSubTab === 'auto'}
           <Sparkles class="w-5 h-5" />
         {:else}
           <ListMusic class={isCollapsed ? "w-5 h-5" : "w-4 h-4"} />
@@ -174,7 +175,7 @@
         <div class="pl-4 pr-1 py-0.5 space-y-0 border-l-2 border-brand-accent/30 ml-4 my-0.5">
           <button
             onclick={() => openPlaylistsSubTab("auto")}
-            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {collectionStore.activeTab === 'playlists' && collectionStore.playlistsSubTab === 'auto' && !collectionStore.selectedPlaylistId && !collectionStore.selectedAutoPlaylist ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
+            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {navigationStore.activeTab === 'playlists' && navigationStore.playlistsSubTab === 'auto' && !navigationStore.selectedPlaylistId && !navigationStore.selectedAutoPlaylist ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
           >
             <div class="flex items-center gap-2 truncate">
               <Sparkles class="w-3.5 h-3.5" />
@@ -187,7 +188,7 @@
 
           <button
             onclick={() => openPlaylistsSubTab("custom")}
-            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {collectionStore.activeTab === 'playlists' && collectionStore.playlistsSubTab === 'custom' && !collectionStore.selectedPlaylistId && !collectionStore.selectedAutoPlaylist ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
+            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {navigationStore.activeTab === 'playlists' && navigationStore.playlistsSubTab === 'custom' && !navigationStore.selectedPlaylistId && !navigationStore.selectedAutoPlaylist ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
           >
             <div class="flex items-center gap-2 truncate">
               <ListMusic class="w-3.5 h-3.5" />
@@ -202,9 +203,9 @@
             onclick={async () => {
               const queuePl = await playlistsStore.requireQueue();
               playlistsStore.selectPlaylist(queuePl.id);
-              collectionStore.viewPlaylist(queuePl.id);
+              navigationStore.viewPlaylist(queuePl.id);
             }}
-            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {collectionStore.activeTab === 'playlists' && collectionStore.selectedPlaylistId && playlistsStore.playlists.find((p) => p.id === collectionStore.selectedPlaylistId)?.name?.toLowerCase() === 'queue' ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
+            class="w-full flex items-center justify-between px-2.5 py-1 rounded-md text-xs transition-colors {navigationStore.activeTab === 'playlists' && navigationStore.selectedPlaylistId && playlistsStore.playlists.find((p) => p.id === navigationStore.selectedPlaylistId)?.name?.toLowerCase() === 'queue' ? 'bg-brand-accent/20 text-brand-accent-text font-semibold' : 'text-brand-text-secondary hover:text-brand-text-primary hover:bg-brand-accent/10'}"
           >
             <div class="flex items-center gap-2 truncate">
               <Layers class="w-3.5 h-3.5" />
@@ -219,8 +220,8 @@
     </div>
 
     <button
-      onclick={() => { collectionStore.activeTab = "lyrics"; }}
-      class="flex items-center gap-3 transition-all duration-150 {collectionStore.activeTab === 'lyrics' ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
+      onclick={() => { navigationStore.activeTab = "lyrics"; }}
+      class="flex items-center gap-3 transition-all duration-150 {navigationStore.activeTab === 'lyrics' ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
       title={i18n.t('sidebar.lyrics')}
     >
       <FileText class={isCollapsed ? "w-5 h-5" : "w-4 h-4"} />
@@ -231,8 +232,8 @@
     {/if}
 
     <button
-      onclick={() => { collectionStore.activeTab = "settings"; }}
-      class="relative flex items-center gap-3 transition-all duration-150 {collectionStore.activeTab === 'settings' ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
+      onclick={() => { navigationStore.activeTab = "settings"; }}
+      class="relative flex items-center gap-3 transition-all duration-150 {navigationStore.activeTab === 'settings' ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
       title={showUpdateBadge ? `${i18n.t('sidebar.settings')} (${i18n.t('settings.updateAvailable', {}, 'Update available')})` : i18n.t('sidebar.settings')}
     >
       <Settings class={isCollapsed ? "w-5 h-5" : "w-4 h-4"} />
@@ -252,8 +253,8 @@
     </button>
 
     <button
-      onclick={() => { collectionStore.activeTab = "help"; }}
-      class="flex items-center gap-3 transition-all duration-150 {collectionStore.activeTab === 'help' ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
+      onclick={() => { navigationStore.activeTab = "help"; }}
+      class="flex items-center gap-3 transition-all duration-150 {navigationStore.activeTab === 'help' ? 'bg-brand-accent text-brand-accent-contrast shadow-lg shadow-brand-accent/20' : 'text-brand-text-secondary hover:bg-brand-accent/10 hover:text-brand-accent-text-hover'} {isCollapsed ? 'justify-center w-10 h-10 rounded-xl p-0' : 'w-full px-3 py-1.5 rounded-lg text-sm font-medium'}"
       title={i18n.t('sidebar.help')}
     >
       <HelpCircle class={isCollapsed ? "w-5 h-5" : "w-4 h-4"} />
