+++ src/lib/components/TopNavigation.test.ts
@@ -3,6 +3,7 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import { render, fireEvent } from "@testing-library/svelte";
 import TopNavigation from "./TopNavigation.svelte";
 import { collectionStore } from "../stores/collection.svelte";
+import { windowLayoutStore } from "../stores/windowLayout.svelte";
 
 vi.mock("@tauri-apps/api/core", () => ({
   invoke: vi.fn().mockResolvedValue([]),
@@ -11,45 +12,45 @@ vi.mock("@tauri-apps/api/core", () => ({
 describe("TopNavigation.svelte", () => {
   beforeEach(() => {
     vi.clearAllMocks();
-    collectionStore.immersiveMode = false;
-    collectionStore.sidebarOpen = true;
-    collectionStore.rightPanelOpen = false;
-    collectionStore.viewportWidth = 1280;
+    windowLayoutStore.immersiveMode = false;
+    windowLayoutStore.sidebarOpen = true;
+    windowLayoutStore.rightPanelOpen = false;
+    windowLayoutStore.viewportWidth = 1280;
   });
 
   it("toggles sidebar compact state when clicking the hamburger menu button", async () => {
-    collectionStore.sidebarOpen = true;
-    collectionStore.setSidebarWidth(256);
+    windowLayoutStore.sidebarOpen = true;
+    windowLayoutStore.setSidebarWidth(256);
     const { getByTitle } = render(TopNavigation);
     const hamburgerBtn = getByTitle("Toggle sidebar (compact / expanded)");
 
     await fireEvent.click(hamburgerBtn);
-    expect(collectionStore.sidebarWidth).toBe(64);
+    expect(windowLayoutStore.sidebarWidth).toBe(64);
 
     await fireEvent.click(hamburgerBtn);
-    expect(collectionStore.sidebarWidth).toBe(256);
+    expect(windowLayoutStore.sidebarWidth).toBe(256);
   });
 
   it("hides the right-panel toggle at the same breakpoint that auto-hides the panel itself", () => {
-    collectionStore.viewportWidth = 1280;
-    expect(collectionStore.isRightPanelAutoHidden).toBe(false);
+    windowLayoutStore.viewportWidth = 1280;
+    expect(windowLayoutStore.isRightPanelAutoHidden).toBe(false);
     const { queryByTitle } = render(TopNavigation);
     expect(queryByTitle("Show Info Panel")).not.toBeNull();
 
-    collectionStore.viewportWidth = 800;
-    expect(collectionStore.isRightPanelAutoHidden).toBe(true);
+    windowLayoutStore.viewportWidth = 800;
+    expect(windowLayoutStore.isRightPanelAutoHidden).toBe(true);
     const { queryByTitle: queryByTitleNarrow } = render(TopNavigation);
     expect(queryByTitleNarrow("Show Info Panel")).toBeNull();
   });
 
   it("hides the sidebar compact/expand toggle at the same breakpoint that auto-collapses it", () => {
-    collectionStore.viewportWidth = 1280;
-    expect(collectionStore.isSidebarAutoCollapsed).toBe(false);
+    windowLayoutStore.viewportWidth = 1280;
+    expect(windowLayoutStore.isSidebarAutoCollapsed).toBe(false);
     const { queryByTitle } = render(TopNavigation);
     expect(queryByTitle("Toggle sidebar (compact / expanded)")).not.toBeNull();
 
-    collectionStore.viewportWidth = 800;
-    expect(collectionStore.isSidebarAutoCollapsed).toBe(true);
+    windowLayoutStore.viewportWidth = 800;
+    expect(windowLayoutStore.isSidebarAutoCollapsed).toBe(true);
     const { queryByTitle: queryByTitleNarrow } = render(TopNavigation);
     expect(queryByTitleNarrow("Toggle sidebar (compact / expanded)")).toBeNull();
   });
